pyproject.toml
[project]
name = "hands-on-large-language-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "colorama>=0.4.6",
    "datasets>=4.8.5",
    "gensim>=4.4.0",
    "hdbscan>=0.8.43",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.9",
    "openai>=2.36.0",
    "pandas>=3.0.3",
    "sentence-transformers>=5.5.0",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
    "transformers>=5.8.0",
    "umap-learn>=0.5.12",
]
